removed secrets and cleaned pycache

+++ app.py
@@ -1,73 +1,68 @@
 from flask import Flask, render_template, request, redirect, url_for
 from flask_sqlalchemy import SQLAlchemy
+from textblob import TextBlob
 import os
+
 app = Flask(__name__)
-app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://{}:{}@{}/{}'.format(
-    os.getenv('DB_USER', 'root'),
-    os.getenv('DB_PASSWORD', '12345'),
-    os.getenv('DB_HOST', 'mysql'),
-    os.getenv('DB_NAME', 'flask')
-)
+
+
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://flask:password@127.0.0.1/flask'
 db = SQLAlchemy(app)
-app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://flask:password@mysql/flask'
 
-
-# Run server:
-# First method: python app.py (app.run needs to be included, like the if statement below)
-# Second method: flask run (after exporting 2 env variables:
-# export FLASK_ENV=development, export FLASK_APP=app.py)
 class Todo(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(100))
     complete = db.Column(db.Boolean)
+    motivation = db.Column(db.String(500))
 
-# create the DB on demand
-@app.before_first_request
-def create_tables():
+with app.app_context():
     db.create_all()
 
 @app.route('/', methods=["GET"])
 def index():
-    print("index")
     t = Todo.query.all()
     return render_template("index.html", list_todo=t)
 
 @app.route('/add', methods=["POST"])
 def add():
-    if request.is_json:
-        data = request.get_json()
-        title = data.get("title")
-    else:
-        title = request.form.get("title")
-    
+    title = request.form.get("title")
     if not title:
-        return {"error": "no title"}, 400
+        return redirect(url_for("index"))
 
-    new_todo = Todo(title=title, complete=False)
+
+    blob = TextBlob(title)
+    sentiment = blob.sentiment.polarity 
+    
+
+    if sentiment > 0.3:
+        motivation_text = "נשמע כמו משימה חיובית! תמשיך עם האנרגיה הזו 🌟"
+    elif sentiment < -0.1:
+        motivation_text = "זו נראית משימה מאתגרת, אבל קטן עליך! 💪"
+    elif len(title) > 20:
+        motivation_text = "וואו, זו משימה מפורטת. בוא נפרק אותה לצעדים קטנים 🧩"
+    else:
+        motivation_text = "עוד צעד בדרך ליעד שלך. קדימה! 🚀"
+
+    new_todo = Todo(title=title, complete=False, motivation=motivation_text)
     db.session.add(new_todo)
     db.session.commit()
-    
-    if request.is_json:
-        return {"message": "added successfully", "title": title}, 200
     return redirect(url_for("index"))
-
 
 @app.route('/update/<int:todo_id>')
 def update(todo_id):
-    todo = Todo.query.filter_by(id=todo_id).first()
-    todo.complete = not todo.complete
-    db.session.commit()
+    todo = Todo.query.get(todo_id)
+    if todo:
+        todo.complete = not todo.complete
+        db.session.commit()
     return redirect(url_for("index"))
 
 @app.route('/delete/<int:todo_id>')
 def delete(todo_id):
-    todo = Todo.query.filter_by(id=todo_id).first()
-    db.session.delete(todo)
-    db.session.commit()
+    todo = Todo.query.get(todo_id)
+    if todo:
+        db.session.delete(todo)
+        db.session.commit()
     return redirect(url_for("index"))
 
 if __name__ == "__main__":
-    #db.create_all()
-    app.run(host=os.getenv('IP', '0.0.0.0'), debug=True)
-    # app.run(host=os.getenv('IP', '0.0.0.0'), debug=True,
-    #         port=int(os.getenv('PORT', 4444)))
+    app.run(host='0.0.0.0', debug=True)
